Remove tolerance constants from the constants module
The motivation behind is to use default values in the individual functions, instead of relying on the generalised ones for the whole algorithm. Reliable generalisation of the numerical tolerance requires too much time and development, than reasonable to do for me at the moment. The constants should be in agreement with the ones of the published algorithm's papers (i.e. niggli, lepage).