Fix a tricky bug
MCLC and ORCC lattices have to stay C-centered. Thus, when conventional cell of SC is constructed one should be carefull not to move the a3. In SC paper the condition b <= c for MCLC is impossible to fulfill while keeping the lattice C-centered. Thus it is not checked for MCLC, but only for MCL lattices.